Create Graphics for Owner

+++ backend/src/main/java/com/embraer/backend/analytics/dto/ListGraphicsOwnersDTO.java
@@ -7,7 +7,7 @@ public class ListGraphicsOwnersDTO {
 
     Long owner_id;
 
-    Integer data_owner;
+    Long data_owner;
 
     String owner_name;
